refactor(BA-6260): use flat userId filter on keypairs instead of nested user filter
The keypair owner is a direct column on the keypair row, so per the
established filter convention (cf. UserFilter.uuid) it should be a flat
`userId` filter, not a nested filter wrapping a related-table attribute.
Nested filters are reserved for cross-table conditions implemented via
EXISTS (e.g. KeypairResourcePolicyFilter.keypair).

- Replace KeypairFilter.user (KeypairUserNestedFilter) with flat user_id: UUIDFilter
- Drop KeypairUserNestedFilter / KeypairUserNestedFilterGQL DTO and GQL types
- Simplify UserAdapter._convert_keypair_filter to a plain column condition
- Regenerate v2-schema.graphql and supergraph.graphql
- Rename test_nested_filters.py to test_filters.py and assert flat user_id
