+++ changes/11896.feature.md
@@ -1,1 +1,1 @@
-Add a user nested filter to the `keypairs` connection on `adminKeypairResourcePoliciesV2` for filtering a policy's keypairs by their owner.
+Add a `userId` filter to the `keypairs` connection on `adminKeypairResourcePoliciesV2` for filtering a policy's keypairs by their owner.
