+++ docs/manager/graphql-reference/v2-schema.graphql
@@ -5555,10 +5555,8 @@ input KeypairFilter {
   accessKey: StringFilter = null
   resourcePolicy: StringFilter = null
 
-  """
-  Added in UNRELEASED. Filter keypairs by their owner. A keypair matches if its owner satisfies the conditions.
-  """
-  user: KeypairUserNestedFilter = null
+  """Added in UNRELEASED. Filter keypairs by the UUID of their owner."""
+  userId: UUIDFilter = null
   createdAt: DateTimeFilter = null
   lastUsed: DateTimeFilter = null
   AND: [KeypairFilter!] = null
@@ -5706,13 +5704,6 @@ enum KeypairResourcePolicyV2OrderField {
   IDLE_TIMEOUT
   MAX_CONCURRENT_SFTP_SESSIONS
   MAX_PENDING_SESSION_COUNT
-}
-
-"""
-Added in UNRELEASED. Nested filter for the user that owns a keypair. Filters keypairs to those owned by a user matching the specified conditions.
-"""
-input KeypairUserNestedFilter {
-  userId: UUIDFilter = null
 }
 
 """Added in 26.4.2. Input for listing files in a virtual folder."""
