+++ src/ai/backend/common/dto/manager/v2/keypair/__init__.py
@@ -12,7 +12,6 @@ from ai.backend.common.dto.manager.v2.keypair.request import (
     AdminUpdateKeypairInput,
     KeypairFilter,
     KeypairOrderBy,
-    KeypairUserNestedFilter,
     RevokeMyKeypairInput,
     SearchKeypairsRequest,
     SearchMyKeypairsRequest,
@@ -57,7 +56,6 @@ __all__ = (
     "KeypairNode",
     "KeypairOrderBy",
     "KeypairOrderField",
-    "KeypairUserNestedFilter",
     "RevokeMyKeypairInput",
     "RevokeMyKeypairPayload",
     "SearchKeypairsRequest",
