+++ src/ai/backend/common/dto/manager/v2/keypair/request.py
@@ -25,23 +25,11 @@ __all__ = (
     "AdminUpdateKeypairInput",
     "KeypairFilter",
     "KeypairOrderBy",
-    "KeypairUserNestedFilter",
     "RevokeMyKeypairInput",
     "SearchMyKeypairsRequest",
     "SwitchMyMainAccessKeyInput",
     "UpdateMyKeypairInput",
 )
-
-
-class KeypairUserNestedFilter(BaseRequestModel):
-    """Nested filter for the user that owns a keypair.
-
-    Filters keypairs to those owned by a user matching the specified conditions.
-    """
-
-    user_id: UUIDFilter | None = Field(
-        default=None, description="Filter by the UUID of the keypair owner."
-    )
 
 
 class KeypairFilter(BaseRequestModel):
@@ -51,12 +39,9 @@ class KeypairFilter(BaseRequestModel):
     is_admin: bool | None = None
     access_key: StringFilter | None = None
     resource_policy: StringFilter | None = None
-    user: KeypairUserNestedFilter | None = Field(
+    user_id: UUIDFilter | None = Field(
         default=None,
-        description=(
-            "Nested filter on the keypair's owner. "
-            "Matches keypairs owned by a user satisfying the conditions."
-        ),
+        description="Filter by the UUID of the keypair owner.",
     )
     created_at: DateTimeFilter | None = None
     last_used: DateTimeFilter | None = None
