+++ src/ai/backend/manager/api/gql/keypair/types/filters.py
@@ -11,9 +11,6 @@ from ai.backend.common.dto.manager.v2.keypair.request import (
 from ai.backend.common.dto.manager.v2.keypair.request import (
     KeypairOrderBy as KeypairOrderByDTO,
 )
-from ai.backend.common.dto.manager.v2.keypair.request import (
-    KeypairUserNestedFilter as KeypairUserNestedFilterDTO,
-)
 from ai.backend.common.meta.meta import NEXT_RELEASE_VERSION
 from ai.backend.manager.api.gql.base import (
     DateTimeFilter,
@@ -28,22 +25,6 @@ from ai.backend.manager.api.gql.decorators import (
     gql_pydantic_input,
 )
 from ai.backend.manager.api.gql.pydantic_compat import PydanticInputMixin
-
-
-@gql_pydantic_input(
-    BackendAIGQLMeta(
-        added_version=NEXT_RELEASE_VERSION,
-        description=(
-            "Nested filter for the user that owns a keypair. "
-            "Filters keypairs to those owned by a user matching the specified conditions."
-        ),
-    ),
-    name="KeypairUserNestedFilter",
-)
-class KeypairUserNestedFilterGQL(PydanticInputMixin[KeypairUserNestedFilterDTO]):
-    """Nested filter for the keypair owner."""
-
-    user_id: UUIDFilter | None = None
 
 
 @gql_pydantic_input(
@@ -64,13 +45,10 @@ class KeypairFilterGQL(PydanticInputMixin[KeypairFilterDTO]):
     is_admin: bool | None = None
     access_key: StringFilter | None = None
     resource_policy: StringFilter | None = None
-    user: KeypairUserNestedFilterGQL | None = gql_added_field(
+    user_id: UUIDFilter | None = gql_added_field(
         BackendAIGQLMeta(
             added_version=NEXT_RELEASE_VERSION,
-            description=(
-                "Filter keypairs by their owner. "
-                "A keypair matches if its owner satisfies the conditions."
-            ),
+            description="Filter keypairs by the UUID of their owner.",
         ),
         default=None,
     )
